requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
indic-transliteration
opencv-python-headless
pyspellchecker
Pillow 